% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ral_colors}
\alias{ral_colors}
\title{RAL color vector index by RAL<number>}
\format{A vector with 213 colors}
\source{
\url{https://en.wikipedia.org/wiki/List_of_RAL_colors}
}
\usage{
ral_colors
}
\description{
RAL color vector index by RAL<number>
}
\keyword{datasets}

