% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ralcolors}
\alias{ralcolors}
\title{RAL color vector indexed by RAL<number>}
\format{A vector with 213 colors}
\source{
\url{https://en.wikipedia.org/wiki/List_of_RAL_colors}
}
\usage{
ralcolors
}
\description{
RAL color vector indexed by RAL<number>
}
\keyword{datasets}
